Upgrade to Traefik v2

+++ init.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 echo "[$0] Initializing..."
-sudo mkdir /opt/traefik
-sudo touch /opt/traefik/acme.json && sudo chmod 600 /opt/traefik/acme.json
-cp .env.sample .env
-echo "[$0] Please edit .env file"
+docker network create traefik-network 2&>1 || true
+if [[ ! -f .env ]]; then
+  cp .env.sample .env
+  echo "[$0] Please edit .env file"
+fi
+echo "[$0] Done."
 exit 0
